More normalization of the filters so it's not always assumed as Series.
