+++ Kavita.Database.Tests/Repositories/PersonRepositoryTests.cs
@@ -5,8 +5,8 @@ using System.Linq;
 using System.Threading.Tasks;
 using Kavita.Common.Helpers;
 using Kavita.Models.Builders;
+using Kavita.Models.DTOs.Filtering.v2.Requests;
 using Kavita.Models.DTOs.Metadata.Browse;
-using Kavita.Models.DTOs.Metadata.Browse.Requests;
 using Kavita.Models.Entities;
 using Kavita.Models.Entities.Enums;
 using Kavita.Models.Entities.Person;
