+++ Kavita.Database/Repositories/AnnotationRepository.cs
@@ -14,7 +14,7 @@ using Kavita.Models.DTOs;
 using Kavita.Models.DTOs.Annotations;
 using Kavita.Models.DTOs.Filtering.v2;
 using Kavita.Models.DTOs.Filtering.v2.FilterFields;
-using Kavita.Models.DTOs.Metadata.Browse.Requests;
+using Kavita.Models.DTOs.Filtering.v2.Requests;
 using Kavita.Models.DTOs.Reader;
 using Kavita.Models.Entities.User;
 using Microsoft.EntityFrameworkCore;
