+++ Kavita.Database/Repositories/PersonRepository.cs
@@ -14,8 +14,8 @@ using Kavita.Database.Extensions.Filters;
 using Kavita.Models.DTOs;
 using Kavita.Models.DTOs.Filtering.v2;
 using Kavita.Models.DTOs.Filtering.v2.FilterFields;
+using Kavita.Models.DTOs.Filtering.v2.Requests;
 using Kavita.Models.DTOs.Metadata.Browse;
-using Kavita.Models.DTOs.Metadata.Browse.Requests;
 using Kavita.Models.DTOs.Person;
 using Kavita.Models.Entities;
 using Kavita.Models.Entities.Enums;
