+++ Kavita.Models/DTOs/Filtering/v2/Requests/ReadingListFilterDto.cs
@@ -1,11 +1,10 @@
-#nullable enable
-using System.Collections.Generic;
-using Kavita.Models.DTOs.Filtering.v2;
+﻿using System.Collections.Generic;
 using Kavita.Models.DTOs.Filtering.v2.SortOptions;
 
-namespace Kavita.Models.DTOs.Metadata.Browse.Requests;
+namespace Kavita.Models.DTOs.Filtering.v2.Requests;
+#nullable enable
 
-public class AnnotationFilterDto : IFilterDto<AnnotationFilterStatementDto>
+public sealed record ReadingListFilterDto : IFilterDto<ReadingListFilterStatementDto>
 {
     /// <summary>
     /// Not used - For parity with Series Filter
@@ -15,9 +14,9 @@ public class AnnotationFilterDto : IFilterDto<AnnotationFilterStatementDto>
     /// Not used - For parity with Series Filter
     /// </summary>
     public string? Name { get; set; }
-    public ICollection<AnnotationFilterStatementDto> Statements { get; set; } = [];
+    public ICollection<ReadingListFilterStatementDto> Statements { get; set; } = [];
     public FilterCombination Combination { get; set; } = FilterCombination.And;
-    public AnnotationSortOptionDto? SortOptions { get; set; }
+    public ReadingListSortOptionDto? SortOptions { get; set; }
 
     /// <summary>
     /// Limit the number of rows returned. Defaults to not applying a limit (aka 0)
