+++ Kavita.Models/DTOs/Filtering/v2/Requests/SeriesFilterV2Dto.cs
@@ -1,13 +1,13 @@
 ﻿using System.Collections.Generic;
 using Kavita.Models.DTOs.Filtering.v2.SortOptions;
 
-namespace Kavita.Models.DTOs.Filtering.v2;
+namespace Kavita.Models.DTOs.Filtering.v2.Requests;
 #nullable enable
 
 /// <summary>
 /// Metadata filtering for v2 API only
 /// </summary>
-public sealed record SeriesFilterV2Dto : IFilterDto<FilterStatementDto>
+public sealed record SeriesFilterV2Dto : IFilterDto<SeriesFilterStatementDto>
 {
     /// <summary>
     /// Not used in the UI.
@@ -17,7 +17,7 @@ public sealed record SeriesFilterV2Dto : IFilterDto<FilterStatementDto>
     /// The name of the filter
     /// </summary>
     public string? Name { get; set; }
-    public ICollection<FilterStatementDto> Statements { get; set; } = [];
+    public ICollection<SeriesFilterStatementDto> Statements { get; set; } = [];
     public FilterCombination Combination { get; set; } = FilterCombination.And;
     public SeriesSortOptionDto? SortOptions { get; set; }
